Break runBot blocks out into their own methods
I moved the block that checked to see if the rfidBucket existed into the
config method, then I moved the block that checked members.ps1.org for
the rfid tag code into its own method. I cleaned up the logic in the
runBot method a good bit and added a lot of logging.